Reset VCR cassettes, moved viewing party tests into their own folder
