+++ spec/fixtures/vcr_cassettes/UserFacade/facade_returns_details/returns_a_movie_title_from_TMDB.yml
@@ -24,8 +24,6 @@ http_interactions:
       - chunked
       Connection:
       - keep-alive
-      Date:
-      - Sat, 02 Dec 2023 03:09:10 GMT
       Server:
       - openresty
       X-Memc:
@@ -36,21 +34,23 @@ http_interactions:
       - '8441'
       X-Memc-Expires:
       - '14023'
+      Date:
+      - Sat, 02 Dec 2023 04:10:22 GMT
       Etag:
       - W/"441cdda54d5e4a2e9f5161a25a748ced"
       Vary:
       - Accept-Encoding
       - Origin
       X-Cache:
-      - Miss from cloudfront
+      - RefreshHit from cloudfront
       Via:
-      - 1.1 5257b3c7acd3727a389946b312557ff2.cloudfront.net (CloudFront)
+      - 1.1 025e84e972a461c1dd140ac5fbbbd0aa.cloudfront.net (CloudFront)
       X-Amz-Cf-Pop:
       - DEN52-P2
       Alt-Svc:
       - h3=":443"; ma=86400
       X-Amz-Cf-Id:
-      - 6M3yi7JGt82vOJJvFESXZiTR1G9XKnYUYjnKVJKok6nxKW5QrhwQKw==
+      - dltZnwn6Gu8z3w45N3aC7WvFkSStsD7ymNgAxWXJN_NtvZjtPxAgjg==
     body:
       encoding: ASCII-8BIT
       string: '{"adult":false,"backdrop_path":"/ctMserH8g2SeOAnCw5gFjdQF8mo.jpg","belongs_to_collection":null,"budget":145000000,"genres":[{"id":35,"name":"Comedy"},{"id":12,"name":"Adventure"},{"id":14,"name":"Fantasy"}],"homepage":"https://www.barbie-themovie.com","id":346698,"imdb_id":"tt1517268","original_language":"en","original_title":"Barbie","overview":"Barbie
@@ -63,5 +63,5 @@ http_interactions:
         Bros. Pictures","origin_country":"US"}],"production_countries":[{"iso_3166_1":"GB","name":"United
         Kingdom"},{"iso_3166_1":"US","name":"United States of America"}],"release_date":"2023-07-19","revenue":1441800000,"runtime":114,"spoken_languages":[{"english_name":"English","iso_639_1":"en","name":"English"}],"status":"Released","tagline":"She''s
         everything. He''s just Ken.","title":"Barbie","video":false,"vote_average":7.2,"vote_count":6092}'
-  recorded_at: Sat, 02 Dec 2023 03:09:10 GMT
+  recorded_at: Sat, 02 Dec 2023 04:10:22 GMT
 recorded_with: VCR 6.2.0
