--- spec/fixtures/vcr_cassettes/spirited_away_details.yml
@@ -1,59 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://api.themoviedb.org/3/movie/219?api_key=<TMDB_API_KEY>&language=en-US
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      User-Agent:
-      - Faraday v2.7.12
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - application/json;charset=utf-8
-      Transfer-Encoding:
-      - chunked
-      Connection:
-      - keep-alive
-      Date:
-      - Thu, 30 Nov 2023 22:59:57 GMT
-      Server:
-      - openresty
-      X-Memc:
-      - HIT
-      X-Memc-Key:
-      - 41ec26e8a91fb3d7b6afac89412b59d4
-      X-Memc-Age:
-      - '11373'
-      X-Memc-Expires:
-      - '10290'
-      Etag:
-      - W/"0299c337da765981d59a0bbf3a879e8e"
-      Vary:
-      - Accept-Encoding
-      - Origin
-      X-Cache:
-      - Miss from cloudfront
-      Via:
-      - 1.1 e5eb6c7f102110906ece71324dc7e664.cloudfront.net (CloudFront)
-      X-Amz-Cf-Pop:
-      - DEN52-P3
-      Alt-Svc:
-      - h3=":443"; ma=86400
-      X-Amz-Cf-Id:
-      - VrrDVnavF85tuzeT5NEsqDyxXBuqFp6KgOkymhiGOCTLHYdSjhAofQ==
-    body:
-      encoding: ASCII-8BIT
-      string: !binary |-
-        eyJhZHVsdCI6ZmFsc2UsImJhY2tkcm9wX3BhdGgiOiIvNmhBOGVOMzdLeThDNlg2dUp3S3NCTlR1ajR0LmpwZyIsImJlbG9uZ3NfdG9fY29sbGVjdGlvbiI6bnVsbCwiYnVkZ2V0IjoxMjg5OTg2NywiZ2VucmVzIjpbeyJpZCI6MzUsIm5hbWUiOiJDb21lZHkifSx7ImlkIjoxOCwibmFtZSI6IkRyYW1hIn0seyJpZCI6MTA3NDksIm5hbWUiOiJSb21hbmNlIn1dLCJob21lcGFnZSI6IiIsImlkIjoyMTksImltZGJfaWQiOiJ0dDA0NDE5MDkiLCJvcmlnaW5hbF9sYW5ndWFnZSI6ImVzIiwib3JpZ2luYWxfdGl0bGUiOiJWb2x2ZXIiLCJvdmVydmlldyI6IldoaWxlIGEgd29tYW4gYXR0ZW1wdHMgdG8gY292ZXIgdXAgaGVyIGRhdWdodGVyJ3MgbXVyZGVyIGFuZCByZWludmVudCBoZXIgbGlmZSwgaGVyIHNpc3RlciBpcyB2aXNpdGVkIGJ5IHRoZWlyIG1vdGhlciwgd2hvIHdhcyB0aG91Z2h0IHRvIGhhdmUgZGllZC4iLCJwb3B1bGFyaXR5IjoyMS41MzUsInBvc3Rlcl9wYXRoIjoiL3ZpeDM3S2M3RVVVUk1DNVJnZDZaNkthRWo1bi5qcGciLCJwcm9kdWN0aW9uX2NvbXBhbmllcyI6W3siaWQiOjQ5LCJsb2dvX3BhdGgiOiIveHBmNWlIZHZ2QnRzSDhqQk1sZ0lKSEFFVDBjLnBuZyIsIm5hbWUiOiJFbCBEZXNlbyIsIm9yaWdpbl9jb3VudHJ5IjoiRVMifSx7ImlkIjo5MzM1LCJsb2dvX3BhdGgiOm51bGwsIm5hbWUiOiJDYW5hbCsgRXNwYcOxYSIsIm9yaWdpbl9jb3VudHJ5IjoiRVMifSx7ImlkIjoyOTI1OCwibG9nb19wYXRoIjpudWxsLCJuYW1lIjoiTWluaXN0ZXJpbyBkZSBDdWx0dXJhIEVzcGHDsWEiLCJvcmlnaW5fY291bnRyeSI6IiJ9LHsiaWQiOjk4MSwibG9nb19wYXRoIjoiL3BrcmRPOXlrcGxPNFpjOGVkaWtvU1g1QVBVTi5wbmciLCJuYW1lIjoiVFZFIiwib3JpZ2luX2NvdW50cnkiOiJFUyJ9LHsiaWQiOjExNjUyLCJsb2dvX3BhdGgiOm51bGwsIm5hbWUiOiJXYXJuZXIgQnJvcy4gRW50ZXJ0YWlubWVudCBFc3Bhw7FhIiwib3JpZ2luX2NvdW50cnkiOiJFUyJ9LHsiaWQiOjU4LCJsb2dvX3BhdGgiOiIvdm9ZQ3dsQkhKUUFOdGp2bTVNTklrQ0YxZERILnBuZyIsIm5hbWUiOiJTb255IFBpY3R1cmVzIENsYXNzaWNzIiwib3JpZ2luX2NvdW50cnkiOiJVUyJ9LHsiaWQiOjEwMTQ2LCJsb2dvX3BhdGgiOiIveG5GSU9lcTVjS3cwOWtDV3FWN2ZvV0RlNEFBLnBuZyIsIm5hbWUiOiJGb2N1cyBGZWF0dXJlcyIsIm9yaWdpbl9jb3VudHJ5IjoiVVMifV0sInByb2R1Y3Rpb25fY291bnRyaWVzIjpbeyJpc29fMzE2Nl8xIjoiRVMiLCJuYW1lIjoiU3BhaW4ifV0sInJlbGVhc2VfZGF0ZSI6IjIwMDYtMDMtMTciLCJyZXZlbnVlIjo4NTU4MjQwNywicnVudGltZSI6MTIxLCJzcG9rZW5fbGFuZ3VhZ2VzIjpbeyJlbmdsaXNoX25hbWUiOiJTcGFuaXNoIiwiaXNvXzYzOV8xIjoiZXMiLCJuYW1lIjoiRXNwYcOxb2wifV0sInN0YXR1cyI6IlJlbGVhc2VkIiwidGFnbGluZSI6IlJldHVybi4iLCJ0aXRsZSI6IlZvbHZlciIsInZpZGVvIjpmYWxzZSwidm90ZV9hdmVyYWdlIjo3LjQ2OCwidm90ZV9jb3VudCI6MTU4N30=
-  recorded_at: Thu, 30 Nov 2023 22:59:57 GMT
-recorded_with: VCR 6.2.0
